implement attach command and server background thread

+++ src/main/java/socs/network/node/RouterDescription.java
@@ -3,7 +3,7 @@ package socs.network.node;
 public class RouterDescription {
   //used to socket communication
   String processIPAddress;
-  short processPortNumber;
+  int processPortNumber;
   //used to identify the router in the simulated network space
   String simulatedIPAddress;
   //status of the router
